+++ src/refactoring/App.tsx
@@ -2,7 +2,6 @@ import { useCallback, useState } from "react";
 
 import { CartPage } from "./pages/cart/CartPage";
 import { AdminPage } from "./pages/admin/AdminPage";
-import { useCoupons, useProducts } from "./hooks";
 import Nav from "./components/layouts/Nav";
 import Layout from "./components/layouts/Layout";
 
@@ -12,26 +11,11 @@ const App = () => {
     setIsAdmin((prev) => !prev);
   }, []);
 
-  const { products, updateProduct, addProduct } = useProducts();
-  const { coupons, addCoupon } = useCoupons();
-
   return (
     <Layout>
       <Nav isAdmin={isAdmin} onToggleAdmin={onToggleAdmin} />
 
-      <main className="container mx-auto mt-6">
-        {isAdmin ? (
-          <AdminPage
-            products={products}
-            coupons={coupons}
-            onProductUpdate={updateProduct}
-            onProductAdd={addProduct}
-            onCouponAdd={addCoupon}
-          />
-        ) : (
-          <CartPage />
-        )}
-      </main>
+      <main className="container mx-auto mt-6">{isAdmin ? <AdminPage /> : <CartPage />}</main>
     </Layout>
   );
 };
